test/web-platform-tests/to-upstream/css/selectors/i18n/css3-selectors-lang-005-dont-upstream.html

<!DOCTYPE html>
<!-- jsdom only test without offsetWidth -->
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>:lang(es), lang="es-MX"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#lang-pseudo'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
#colonlangcontroltest:lang(xx) { display:none; }
#box:lang(es) { width: 100px; }
</style>
</head>
<body>

<div class="test"><div id="box" lang="es-MX">&#xA0;</div></div>
<p lang='xx' id='colonlangcontroltest'>This test failed because it relies on :lang for results, but :lang is not supported by this browser.

<!--Notes:
This tests a detail related to :lang support. If :lang is not supported, a message will appear and the test will fail.
-->
<script>
"use strict";
test(() => {
  assert_equals(window.getComputedStyle(document.getElementById("colonlangcontroltest")).display, "none");
  assert_true(document.getElementById("box").matches("#box:lang(es)"));
}, "A :lang value will match a lang attribute value when the latter contains additional subtags.");
</script>

<div id='log'></div>

</body>
</html>
